apply feedback

+++ src/main/java/fr/formationacademy/scpiinvestplusapi/dto/LocationRequest.java
@@ -3,6 +3,8 @@ package fr.formationacademy.scpiinvestplusapi.dto;
 import jakarta.validation.constraints.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class LocationRequest {
     private String country;
 
     @PositiveOrZero(message = "Country percentage must be non-negative")
-    private Float countryPercentage;
+    private BigDecimal countryPercentage;
 
     @NotNull(message = "SCPI ID is required")
     private Integer scpiId;
