+++ src/main/java/fr/formationacademy/scpiinvestplusapi/dto/SectorDtoOut.java
@@ -6,11 +6,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 @Builder
 public class SectorDtoOut {
     private SectorId id;
-    private Float sectorPercentage;
+    private BigDecimal sectorPercentage;
 }
